With this code, the AA Server now uses the AA Agent code to send LLDP PDU's.
This modification to AA-Server sends LLDP PDU's using the AA Agent mechanism, replacing the existing
working LLDPD levent send mechanism. When LLDPD interface layer discovers a port, aatransi_port_create()
is now called to ensure the port is created in the context of the AA Agent code, so that when it's time
to send a pdu, the Agent knows about the hardware interfaces on which to send the packet.
aatransi_packet_compose() now calls LLDPD interfaces_send_helper() if cfg_ext.send is not defined.

removed some warnings as well.