Within lldp_decode, we add newly discovered isid-vlan mappings to the lldpd local hardware port, as well as to the aa agent local hardware port.
This change makes sure we don't add duplicate isid-vlan mappings to the
aa agent local hardware port. This is necessary because periodically the
lldpd local hardware structures are cleared out, and new isid-vlan mappings
are added to both local hardware structures, but we should not add them to
the aa agent unless they're new.